requirements.txt
docopt
numpy
sympy
python-flint
numpy
sympy==1.10
colorama
pycryptodome